Bug Fix - Complex Puzzles
The previous version could not handle puzzles with multiple correct solutions or that were missing many clues. This version fixes those bugs by getting rid of the .includes() call in the rowChecker() function.

In addition, y and x were switched throughout to r and c to reflect better the row and column structure.